All Tests are passing

+++ src/main/java/com/revature/movieapp/repos/FAQs_Repo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FAQs_Repo extends JpaRepository<FAQs,Integer> {
+public interface FAQs_Repo extends JpaRepository<FAQs,Long> {
     //This is also referred to as a Data-Access-Object.
            }
